Move findGoogleDriveAssetsInGraph to google-drive-kit package
