+++ packages/google-drive-kit/src/board-server/utils.ts
@@ -4,9 +4,13 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import type { GraphTag } from "@breadboard-ai/types";
-import type { StoredProperties } from "./operations.js";
+import type {
+  GraphDescriptor,
+  GraphTag,
+  LLMContent,
+} from "@breadboard-ai/types";
 import type { NarrowedDriveFile } from "../google-drive-client.js";
+import type { StoredProperties } from "./operations.js";
 
 /** Delay between GDrive API retries. */
 const RETRY_MS = 200;
@@ -160,3 +164,45 @@ export type AppProperties = {
   tags: GraphTag[];
   thumbnailUrl?: string;
 };
+
+export function findGoogleDriveAssetsInGraph(graph: GraphDescriptor): string[] {
+  // Use a set because there can be duplicates.
+  const fileIds = new Set<string>();
+
+  if (graph.assets) {
+    for (const asset of Object.values(graph.assets)) {
+      // Cast needed because `data` is very broadly typed as `NodeValue`.
+      const firstPart = (asset.data as LLMContent[])[0]?.parts?.[0];
+      if (firstPart) {
+        if ("fileData" in firstPart && asset.metadata?.subType === "gdrive") {
+          const fileId = firstPart.fileData?.fileUri;
+          if (fileId) {
+            fileIds.add(fileId);
+          }
+        }
+        if ("storedData" in firstPart) {
+          const fileId = extractGoogleDriveFileId(firstPart.storedData?.handle);
+          if (fileId) {
+            fileIds.add(fileId);
+          }
+        }
+      }
+    }
+  }
+
+  // Theme splash images are not listed in assets.
+  const themes = graph.metadata?.visual?.presentation?.themes;
+  if (themes) {
+    for (const theme of Object.values(themes)) {
+      const splashHandle = theme.splashScreen?.storedData?.handle;
+      if (splashHandle) {
+        const fileId = extractGoogleDriveFileId(splashHandle);
+        if (fileId) {
+          fileIds.add(fileId);
+        }
+      }
+    }
+  }
+
+  return [...fileIds];
+}
