+++ packages/shared-ui/src/elements/canvas-controller/canvas-controller.ts
@@ -74,7 +74,8 @@ import { emptyStyles } from "../../styles/host/colors-empty.js";
 const SIDE_ITEM_KEY = "bb-canvas-controller-side-nav-item";
 
 import "./empty-state.js";
-import { findGoogleDriveAssetsInGraph, isEmpty } from "../../utils/utils.js";
+import { findGoogleDriveAssetsInGraph } from "@breadboard-ai/google-drive-kit/board-server/utils.js";
+import { isEmpty } from "../../utils/utils.js";
 import type { DriveFileIdWithOptionalResourceKey } from "../google-drive/google-drive-picker.js";
 
 @customElement("bb-canvas-controller")
