+++ packages/shared-ui/src/utils/utils.ts
@@ -16,5 +16,4 @@ export * as YouTube from "./youtube.js";
 export { isCtrlCommand } from "./is-ctrl-command.js";
 export { blankBoard } from "./blank-board.js";
 export { stringifyPermission } from "./stringify-permission.js";
-export { findGoogleDriveAssetsInGraph } from "./find-google-drive-assets-in-graph.js";
 export { isEmpty } from "./graph-is-empty.js";
